Works around crash seen when using layout with AsyncDisplayKit.
AsyncDisplayKit was requesting a layout before the collection view had items (or sections). That led to a crash because the `numberOfItems` computed property was querying the collection view for the number of items in non-existent section 0. This fix updates the `numberOfItems` computed property to always return 0 if there aren’t any sections in the collection view.